created  a usercard for the mentors available.
